# Conflicts:
#	composeApp/src/commonMain/kotlin/org/example/project/ViewSampleCollectionScreen.kt

+++ composeApp/src/commonMain/kotlin/com/dev/database/entity/Entity.kt
@@ -38,7 +38,7 @@ enum class FilterDateOption {
 // sampleForm
 @Serializable
 data class SampleForm(
-    @SerialName("form_id") val formId: Int,
+    @SerialName("form_id") val formId: Long,
     @SerialName("form_name") val formName: String,
     @SerialName("form_active") val formActive: Int
 )
